refactor: removed image shape from Particle.ts
