+++ core/main/src/Core/Particle.ts
@@ -2,7 +2,6 @@ import type { Container } from "./Container";
 import type { IVelocity } from "./Interfaces/IVelocity";
 import type { IParticleValueAnimation } from "./Interfaces/IParticleValueAnimation";
 import type { ICoordinates, ICoordinates3d } from "./Interfaces/ICoordinates";
-import type { IParticleImage } from "./Interfaces/IParticleImage";
 import type { IHsl, IRgb } from "./Interfaces/Colors";
 import type { IShapeValues } from "../Options/Interfaces/Particles/Shape/IShapeValues";
 import type { IBubbleParticleData } from "./Interfaces/IBubbleParticleData";
@@ -20,11 +19,8 @@ import {
     ShapeType,
     StartValueType,
 } from "../Enums";
-import { ImageDrawer } from "../ShapeDrawers/ImageDrawer";
-import type { IImageShape } from "../Options/Interfaces/Particles/Shape/IImageShape";
 import type { RecursivePartial } from "../Types";
-import { colorToHsl, colorToRgb, getHslFromAnimation, NumberUtils, Plugins, replaceColorSvg, Utils } from "../Utils";
-import type { IShapeDrawer } from "./Interfaces/IShapeDrawer";
+import { colorToHsl, colorToRgb, getHslFromAnimation, NumberUtils, Plugins, Utils } from "../Utils";
 import type { IDelta } from "./Interfaces/IDelta";
 import { Mover } from "./Particle/Mover";
 import type { ILink } from "./Interfaces/ILink";
@@ -74,10 +70,10 @@ export class Particle implements IParticle {
     public spinAcceleration?: number;
     public orbitRadius?: number;
     public orbitRotation?: number;
+    public close: boolean;
+    public fill: boolean;
 
-    public readonly close: boolean;
     public readonly direction: MoveDirection | keyof typeof MoveDirection | MoveDirectionAlt;
-    public readonly fill: boolean;
     public readonly infection: IParticleInfection;
     public readonly loops: IParticleLoops;
     public readonly stroke: Stroke;
@@ -92,7 +88,6 @@ export class Particle implements IParticle {
     public readonly orbitColor?: IHsl;
     public readonly velocity: IVelocity;
     public readonly shape: ShapeType | string;
-    public readonly image?: IParticleImage;
     public readonly initialPosition: ICoordinates;
     public readonly initialVelocity: IVelocity;
     public readonly shapeData?: IShapeValues;
@@ -101,7 +96,7 @@ export class Particle implements IParticle {
 
     constructor(
         public readonly id: number,
-        private readonly container: Container,
+        public readonly container: Container,
         position?: ICoordinates,
         overrideOptions?: RecursivePartial<IParticles>,
         public readonly group?: string
@@ -402,19 +397,14 @@ export class Particle implements IParticle {
             }
         }
 
+        if (drawer?.loadShape) {
+            drawer?.loadShape(this);
+        }
+
         const sideCountFunc = drawer?.getSidesCount;
 
         if (sideCountFunc) {
             this.sides = sideCountFunc(this);
-        }
-
-        /* if shape is image */
-        const imageShape = this.loadImageShape(container, drawer);
-
-        if (imageShape) {
-            this.image = imageShape.image;
-            this.fill = imageShape.fill;
-            this.close = imageShape.close;
         }
 
         this.stroke =
@@ -453,14 +443,14 @@ export class Particle implements IParticle {
 
         this.lifeDelay = container.retina.reduceFactor
             ? ((NumberUtils.getValue(lifeOptions.delay) * (lifeOptions.delay.sync ? 1 : Math.random())) /
-                  container.retina.reduceFactor) *
-              1000
+            container.retina.reduceFactor) *
+            1000
             : 0;
         this.lifeDelayTime = 0;
         this.lifeDuration = container.retina.reduceFactor
             ? ((NumberUtils.getValue(lifeOptions.duration) * (lifeOptions.duration.sync ? 1 : Math.random())) /
-                  container.retina.reduceFactor) *
-              1000
+            container.retina.reduceFactor) *
+            1000
             : 0;
         this.lifeTime = 0;
         this.livesRemaining = particlesOptions.life.count;
@@ -574,7 +564,7 @@ export class Particle implements IParticle {
     }
 
     private calcPosition(container: Container, position: ICoordinates | undefined, zIndex: number): ICoordinates3d {
-        for (const [, plugin] of container.plugins) {
+        for (const [ , plugin ] of container.plugins) {
             const pluginPos =
                 plugin.particlePosition !== undefined ? plugin.particlePosition(position, this) : undefined;
 
@@ -673,94 +663,5 @@ export class Particle implements IParticle {
         for (let i = 0; i < rate; i++) {
             this.container.particles.addSplitParticle(this);
         }
-    }
-
-    private loadImageShape(
-        container: Container,
-        drawer?: IShapeDrawer
-    ):
-        | {
-              image: IParticleImage | undefined;
-              fill: boolean;
-              close: boolean;
-          }
-        | undefined {
-        if (!(this.shape === ShapeType.image || this.shape === ShapeType.images)) {
-            return;
-        }
-
-        const imageDrawer = drawer as ImageDrawer;
-        const images = imageDrawer.getImages(container).images;
-        const imageData = this.shapeData as IImageShape;
-        const image = images.find((t) => t.source === imageData.src) ?? images[0];
-        const color = this.getFillColor();
-        let imageRes: IParticleImage;
-
-        if (!image) {
-            return;
-        }
-
-        if (image.svgData !== undefined && imageData.replaceColor && color) {
-            const svgColoredData = replaceColorSvg(image, color, this.opacity.value);
-
-            /* prepare to create img with colored svg */
-            const svg = new Blob([svgColoredData], { type: "image/svg+xml" });
-            const domUrl = URL || window.URL || window.webkitURL || window;
-            const url = domUrl.createObjectURL(svg);
-
-            /* create particle img obj */
-            const img = new Image();
-
-            imageRes = {
-                data: image,
-                loaded: false,
-                ratio: imageData.width / imageData.height,
-                replaceColor: imageData.replaceColor ?? imageData.replace_color,
-                source: imageData.src,
-            };
-
-            img.addEventListener("load", () => {
-                if (this.image) {
-                    this.image.loaded = true;
-                    image.element = img;
-                }
-
-                domUrl.revokeObjectURL(url);
-            });
-
-            img.addEventListener("error", () => {
-                domUrl.revokeObjectURL(url);
-
-                // deepcode ignore PromiseNotCaughtGeneral: catch can be ignored
-                Utils.loadImage(imageData.src).then((img2) => {
-                    if (this.image) {
-                        image.element = img2?.element;
-
-                        this.image.loaded = true;
-                    }
-                });
-            });
-
-            img.src = url;
-        } else {
-            imageRes = {
-                data: image,
-                loaded: true,
-                ratio: imageData.width / imageData.height,
-                replaceColor: imageData.replaceColor ?? imageData.replace_color,
-                source: imageData.src,
-            };
-        }
-        if (!imageRes.ratio) {
-            imageRes.ratio = 1;
-        }
-        const fill = imageData.fill ?? this.fill;
-        const close = imageData.close ?? this.close;
-
-        return {
-            image: imageRes,
-            fill,
-            close,
-        };
     }
 }
