+++ core/main/src/ShapeDrawers/ImageDrawer.ts
@@ -1,15 +1,29 @@
 import type { IShapeDrawer } from "../Core/Interfaces/IShapeDrawer";
 import type { IParticle } from "../Core/Interfaces/IParticle";
-import { Utils } from "../Utils";
+import { replaceColorSvg, Utils } from "../Utils";
 import { ShapeType } from "../Enums";
 import type { IImageShape } from "../Options/Interfaces/Particles/Shape/IImageShape";
 import type { IImage } from "../Core/Interfaces/IImage";
 import type { Container } from "../Core/Container";
+import type { Particle } from "../Core/Particle";
+
+interface IParticleImage {
+    source: string;
+    data: IImage;
+    ratio: number;
+    element?: HTMLImageElement;
+    loaded?: boolean;
+    replaceColor: boolean;
+}
 
 interface ContainerImage {
     id: string;
     images: IImage[];
 }
+
+type IImageParticle = IParticle & {
+    image: IParticleImage;
+};
 
 /**
  * @category Shape Drawers
@@ -97,7 +111,7 @@ export class ImageDrawer implements IShapeDrawer {
             return;
         }
 
-        const image = particle.image;
+        const image = (particle as IImageParticle).image;
         const element = image?.data?.element;
 
         if (!element) {
@@ -120,5 +134,91 @@ export class ImageDrawer implements IShapeDrawer {
         if (!image?.data.svgData || !image?.replaceColor) {
             context.globalAlpha = 1;
         }
+    }
+
+    public loadShape(particle: Particle): void {
+        if (!(particle.shape === ShapeType.image || particle.shape === ShapeType.images)) {
+            return;
+        }
+
+        const container = particle.container;
+        const images = this.getImages(container).images;
+        const imageData = particle.shapeData as IImageShape;
+        const image = images.find((t) => t.source === imageData.src) ?? images[0];
+        const color = particle.getFillColor();
+        let imageRes: IParticleImage;
+
+        if (!image) {
+            return;
+        }
+
+        if (image.svgData !== undefined && imageData.replaceColor && color) {
+            const svgColoredData = replaceColorSvg(image, color, particle.opacity.value);
+
+            /* prepare to create img with colored svg */
+            const svg = new Blob([svgColoredData], { type: "image/svg+xml" });
+            const domUrl = URL || window.URL || window.webkitURL || window;
+            const url = domUrl.createObjectURL(svg);
+
+            /* create particle img obj */
+            const img = new Image();
+
+            imageRes = {
+                data: image,
+                ratio: imageData.width / imageData.height,
+                replaceColor: imageData.replaceColor ?? imageData.replace_color,
+                source: imageData.src,
+            };
+
+            img.addEventListener("load", () => {
+                const pImage = ((particle as unknown) as IImageParticle).image;
+                if (pImage) {
+                    pImage.loaded = true;
+                    image.element = img;
+                }
+
+                domUrl.revokeObjectURL(url);
+            });
+
+            img.addEventListener("error", () => {
+                domUrl.revokeObjectURL(url);
+
+                // deepcode ignore PromiseNotCaughtGeneral: catch can be ignored
+                Utils.loadImage(imageData.src).then((img2) => {
+                    const pImage = ((particle as unknown) as IImageParticle).image;
+
+                    if (pImage) {
+                        image.element = img2?.element;
+
+                        pImage.loaded = true;
+                    }
+                });
+            });
+
+            img.src = url;
+        } else {
+            imageRes = {
+                data: image,
+                loaded: true,
+                ratio: imageData.width / imageData.height,
+                replaceColor: imageData.replaceColor ?? imageData.replace_color,
+                source: imageData.src,
+            };
+        }
+        if (!imageRes.ratio) {
+            imageRes.ratio = 1;
+        }
+        const fill = imageData.fill ?? particle.fill;
+        const close = imageData.close ?? particle.close;
+
+        const imageShape = {
+            image: imageRes,
+            fill,
+            close,
+        };
+
+        ((particle as unknown) as IImageParticle).image = imageShape.image;
+        particle.fill = imageShape.fill;
+        particle.close = imageShape.close;
     }
 }
