+++ core/main/src/Types/ShapeDrawerFunctions.ts
@@ -1,5 +1,6 @@
 import type { IParticle } from "../Core/Interfaces/IParticle";
 import type { Container } from "../Core/Container";
+import type { Particle } from "../Core/Particle";
 
 /**
  * @category Types
@@ -39,3 +40,8 @@ export type ShapeDrawerDestroyFunction = (container: Container) => void;
  * @category Types
  */
 export type ShapeDrawerSidesCountFunction = (particle: IParticle) => number;
+
+/**
+ * @category Types
+ */
+export type ShapeDrawerLoadFunction = (particle: Particle) => void;
